requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-decouple==3.7
pytz==2022.7
sqlparse==0.4.3
static3==0.7.0
tzdata==2022.7
psycopg2-binary==2.9.5
whitenoise==6.3.0
Brotli==1.0.9
gunicorn